% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLWL.R
\name{MLWL}
\alias{MLWL}
\title{Marginal likelihood estimation using the Wang-Landau algorithm}
\usage{
MLWL(num_iterations, num_burnin, target_kernel, surrogate_kernel,
  target_logdensity, surrogate_logdensity, learning_rate,
  flatness_criterion, x0)
}
\arguments{
\item{num_iterations}{Total number of iterations}

\item{num_burnin}{Number of iterations for burn-in}

\item{target_kernel}{A Markov kernel invariant to the target distribution}

\item{surrogate_kernel}{A Markov kernel invariant to the surrogate distribution}

\item{target_logdensity}{A function evaluating the log density of the unnormalized target distribution}

\item{surrogate_logdensity}{A function evaluating the log density of the normalized surrogate distribution}

\item{learning_rate}{Initialization of the learning_rate}

\item{flatness_criterion}{The flatness criterion to adaptly decrease the learning rate.}

\item{x0}{An initializing sample}
}
\value{
The log normalizing constant (marginal likelihood) of the target distribution.
}
\description{
Marginal likelihood estimation using the Wang-Landau algorithm
}
