% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalizing two log constants to sum 1}
\usage{
normalization(target_logZ, surrogate_logZ)
}
\arguments{
\item{target_logZ}{Unnormalized log normalizing constants of the target distribution}

\item{surrogate_logZ}{Unnormalized log normalizing constants of the surrogate distribution}
}
\value{
Two normalized log constants
}
\description{
Normalizing two log constants to sum 1
}
